requirements.txt
-i https://pypi.org/simple
alembic==1.11.1 ; python_version >= '3.7'
aniso8601==9.0.1
blinker==1.6.2 ; python_version >= '3.7'
cachelib==0.10.2 ; python_version >= '3.7'
certifi==2023.7.22 ; python_version >= '3.6'
charset-normalizer==3.2.0 ; python_full_version >= '3.7.0'
click==8.1.6 ; python_version >= '3.7'
flask==2.3.2
flask-cors==4.0.0
flask-migrate==4.0.4
flask-restful==0.3.10
flask-session==0.5.0
flask-sqlalchemy==3.0.5
greenlet==2.0.2 ; platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
gunicorn==21.2.0
idna==3.4 ; python_version >= '3.5'
itsdangerous==2.1.2 ; python_version >= '3.7'
jinja2==3.1.2 ; python_version >= '3.7'
mako==1.2.4 ; python_version >= '3.7'
markupsafe==2.1.3 ; python_version >= '3.7'
packaging==23.1 ; python_version >= '3.7'
psycopg2-binary==2.9.6
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlalchemy==2.0.19 ; python_version >= '3.7'
sqlalchemy-serializer==1.4.1
typing-extensions==4.7.1 ; python_version >= '3.7'
urllib3==2.0.4 ; python_version >= '3.7'
werkzeug==2.3.6 ; python_version >= '3.8'
